requirements.txt
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
python-dotenv
streamlit